Ready for parts and firmware
Far enough along now to determine the size of the perfboard to put this
on (6" x 6") and to start working on a BOM and firmware. My plan now is
to get a little ways in to the firmware and place the order when I'm
sure the firmware won't require hardware design changes.